+++ addons/development_console_dotnet/scripts/Main.cs
@@ -1,4 +1,5 @@
 using Godot;
+using Godot.Collections;
 
 namespace InGameConsole;
 
@@ -9,6 +10,34 @@ public partial class Main : EditorPlugin
     public override void _EnablePlugin()
     {
         AddAutoloadSingleton("GameConsole", GameConsoleUiPath);
+
+        if (!ProjectSettings.HasSetting("input/accept"))
+        {
+            var input = new Dictionary();
+            input.Add("deadzone", 0.5f);
+            input.Add("events", new Array()
+            {
+                new InputEventKey() {Keycode = Key.Enter}
+            });
+            ProjectSettings.SetSetting("input/accept", input);
+            ProjectSettings.Save();
+        }
+        
+        if (!ProjectSettings.HasSetting("input/console"))
+        {
+            var input = new Dictionary();
+            input.Add("deadzone", 0.5f);
+            input.Add("events", new Array()
+            {
+                new InputEventKey() {Keycode = Key.Asciitilde}
+            });
+            ProjectSettings.SetSetting("input/console", input);
+            ProjectSettings.Save();
+        }
+        
+        // This forces the editor to restart,
+        // which is needed for some stupid reason to actually update the input settings
+        EditorInterface.Singleton.RestartEditor();
     }
     
     public override void _DisablePlugin()